lxd/network/network/utils: Removes OVN specific helper functions
Moved to own file.
