+++ lxd/network/network_utils.go
@@ -867,29 +867,6 @@ func randomHwaddr(r *rand.Rand) string {
 	return ret.String()
 }
 
-// OVNInstanceDevicePortAdd adds a logical port to the OVN network's internal switch and returns the logical
-// port name for use linking an OVS port on the integration bridge to the logical switch port.
-func OVNInstanceDevicePortAdd(network Network, instanceID int, deviceName string, mac net.HardwareAddr, ips []net.IP) (openvswitch.OVNSwitchPort, error) {
-	// Check network is of type OVN.
-	n, ok := network.(*ovn)
-	if !ok {
-		return "", fmt.Errorf("Network is not OVN type")
-	}
-
-	return n.instanceDevicePortAdd(instanceID, deviceName, mac, ips)
-}
-
-// OVNInstanceDevicePortDelete deletes a logical port from the OVN network's internal switch.
-func OVNInstanceDevicePortDelete(network Network, instanceID int, deviceName string) error {
-	// Check network is of type OVN.
-	n, ok := network.(*ovn)
-	if !ok {
-		return fmt.Errorf("Network is not OVN type")
-	}
-
-	return n.instanceDevicePortDelete(instanceID, deviceName)
-}
-
 // parseIPRange parses an IP range in the format "start-end" and converts it to a shared.IPRange.
 // If allowedNets are supplied, then each IP in the range is checked that it belongs to at least one of them.
 // IPs in the range can be zero prefixed, e.g. "::1" or "0.0.0.1", however they should not overlap with any
